Only import os_utils when plugin is called
Currently, if a dependency of one of the os_ plugins is missing,
running monitorstack with no args returns (for example):

```
OpenStack plugins require access to the OpenStackSDK. \
    Please install "python-openstacksdk". \
    ERROR: No module named keystoneauth1.loading
```

This commit moves the os_utils import in the os_ plugins to within cli,
similar to what is being done in the kvm module.  This allows the
monitorstack cli to operate normally even if a plugin cannot be loaded.

Change-Id: I2ff1ad8c6365f6592191f09ea9c0d543f81e16ba